优化角色收藏功能
