+++ app/src/main/res/layout/fragment_character_pager.xml
@@ -36,13 +36,29 @@
         android:gravity="center"
         android:textAppearance="@style/TextAppearance.MaterialComponents.Overline" />
 
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/fab_love"
+        style="@style/FabStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:contentDescription="@string/tab_love"
+        android:src="@drawable/ic_loved"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <androidx.appcompat.widget.AppCompatImageView
+        android:id="@+id/btn_love"
+        android:layout_width="@dimen/l_icon"
+        android:layout_height="@dimen/l_icon"
+        android:foreground="?selectableItemBackgroundBorderless" />
+
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/character_pager"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constrainedHeight="true"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/name_ex"
+        app:layout_constraintTop_toBottomOf="@id/character_pic"
         app:layout_constraintVertical_weight="1"
         tools:context=".ui.character.basic.CharacterBasicInfoFragment" />
