--- app/src/main/res/xml/fragment_character_basic_info_scene.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<MotionScene xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <ConstraintSet android:id="@+id/expanded">
-        <Constraint
-            android:id="@id/character_pic"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <Constraint
-            android:id="@id/catah"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/s_dp"
-            android:layout_marginStart="@dimen/s_dp"
-            android:layout_marginEnd="@dimen/s_dp"
-            app:layout_constraintTop_toBottomOf="@id/character_pic" />
-
-        <Constraint
-            android:id="@id/fab_love_cbi"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/character_pic" />
-
-    </ConstraintSet>
-
-    <ConstraintSet android:id="@+id/collapsed">
-        <Constraint
-            android:id="@id/character_pic"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/toolbar_height"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <Constraint
-            android:id="@id/catah"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/s_dp"
-            android:layout_marginEnd="@dimen/s_dp"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintTop_toTopOf="@id/character_pic" />
-
-        <Constraint
-            android:id="@id/fab_love_cbi"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/character_pic" />
-    </ConstraintSet>
-
-    <Transition
-        app:constraintSetEnd="@id/collapsed"
-        app:constraintSetStart="@+id/expanded">
-        <KeyFrameSet>
-            <KeyAttribute
-                android:alpha="1.0"
-                app:framePosition="0"
-                app:motionTarget="@id/name_extra" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="50"
-                app:motionTarget="@id/name_extra" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="100"
-                app:motionTarget="@id/name_extra" />
-            <KeyAttribute
-                android:alpha="1.0"
-                app:framePosition="0"
-                app:motionTarget="@id/fab_love_cbi" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="50"
-                app:motionTarget="@id/fab_love_cbi" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="100"
-                app:motionTarget="@id/fab_love_cbi" />
-            <KeyAttribute
-                android:alpha="1.0"
-                app:framePosition="30"
-                app:motionTarget="@id/character_pic" />
-            <KeyAttribute
-                android:alpha="0.0"
-                app:framePosition="100"
-                app:motionTarget="@id/character_pic" />
-
-        </KeyFrameSet>
-        <OnSwipe
-            app:dragDirection="dragUp"
-            app:touchAnchorId="@id/scroll"
-            app:touchAnchorSide="top" />
-    </Transition>
-</MotionScene>
