+++ app/src/main/res/xml/fragment_character_pager_scene.xml
@@ -19,9 +19,26 @@
             android:id="@id/name_ex"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintStart_toStartOf="@id/name"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
             app:layout_constraintEnd_toEndOf="@id/name"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic" />
+            app:layout_constraintStart_toStartOf="@id/name" />
+
+        <Constraint
+            android:id="@id/fab_love"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/character_pic" />
+
+        <Constraint
+            android:id="@id/btn_love"
+            android:layout_width="@dimen/l_icon"
+            android:layout_height="@dimen/l_icon"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
+            app:layout_constraintEnd_toEndOf="@id/fab_love"
+            app:layout_constraintStart_toStartOf="@id/fab_love"
+            app:layout_constraintTop_toBottomOf="@id/character_pic" />
     </ConstraintSet>
 
     <ConstraintSet android:id="@+id/collapsed">
@@ -41,10 +58,26 @@
             android:id="@id/name_ex"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintStart_toStartOf="@id/name"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
             app:layout_constraintEnd_toEndOf="@id/name"
-            app:layout_constraintBottom_toBottomOf="@id/character_pic" />
+            app:layout_constraintStart_toStartOf="@id/name" />
 
+        <Constraint
+            android:id="@id/fab_love"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="@id/character_pic" />
+
+        <Constraint
+            android:id="@id/btn_love"
+            android:layout_width="@dimen/l_icon"
+            android:layout_height="@dimen/l_icon"
+            app:layout_constraintBottom_toBottomOf="@id/character_pic"
+            app:layout_constraintEnd_toEndOf="@id/fab_love"
+            app:layout_constraintStart_toStartOf="@id/fab_love"
+            app:layout_constraintTop_toTopOf="@id/character_pic" />
     </ConstraintSet>
 
     <Transition
@@ -59,6 +92,7 @@
                 android:alpha="0.0"
                 app:framePosition="100"
                 app:motionTarget="@id/character_pic" />
+
             <KeyAttribute
                 android:alpha="0"
                 app:framePosition="0"
@@ -71,6 +105,7 @@
                 android:alpha="1"
                 app:framePosition="100"
                 app:motionTarget="@id/name" />
+
             <KeyAttribute
                 android:alpha="0"
                 app:framePosition="0"
@@ -84,6 +119,23 @@
                 app:framePosition="100"
                 app:motionTarget="@id/name_ex" />
 
+            <KeyAttribute
+                android:alpha="1.0"
+                app:framePosition="0"
+                app:motionTarget="@id/fab_love" />
+            <KeyAttribute
+                android:alpha="0.0"
+                app:framePosition="100"
+                app:motionTarget="@id/fab_love" />
+
+            <KeyAttribute
+                android:alpha="0"
+                app:framePosition="0"
+                app:motionTarget="@id/btn_love" />
+            <KeyAttribute
+                android:alpha="1"
+                app:framePosition="100"
+                app:motionTarget="@id/btn_love" />
         </KeyFrameSet>
         <OnSwipe
             app:dragDirection="dragUp"
